+++ docs/flutter_service_worker.js
@@ -1,16 +1,17 @@
 'use strict';
 const CACHE_NAME = 'flutter-app-cache';
 const RESOURCES = {
-  "/index.html": "ecc976d5c4d73891d4e9d680f683ccbd",
-"/main.dart.js": "1ca9e48762637f83f83ee8bc2f54f5e1",
-"/icons/Icon-192.png": "ac9a721a12bbc803b44f645561ecb1e1",
-"/icons/Icon-512.png": "96e752610906ba2a93c65f8abe1645f1",
-"/manifest.json": "00e0b69b49487ce4f9ff0c5fac8fda49",
-"/assets/LICENSE": "873745433467312fdfb98ff306e156fe",
-"/assets/AssetManifest.json": "2efbb41d7877d10aac9d091f58ccd7b9",
-"/assets/FontManifest.json": "01700ba55b08a6141f33e168c4a6c22f",
-"/assets/packages/cupertino_icons/assets/CupertinoIcons.ttf": "115e937bb829a890521f72d2e664b632",
-"/assets/fonts/MaterialIcons-Regular.ttf": "56d3ffdef7a25659eab6a68a3fbfaf16"
+  "index.html": "ecc976d5c4d73891d4e9d680f683ccbd",
+"/": "ecc976d5c4d73891d4e9d680f683ccbd",
+"main.dart.js": "c0c1e5550a2296fd2ce45d3cf92fc783",
+"icons/Icon-192.png": "ac9a721a12bbc803b44f645561ecb1e1",
+"icons/Icon-512.png": "96e752610906ba2a93c65f8abe1645f1",
+"manifest.json": "00e0b69b49487ce4f9ff0c5fac8fda49",
+"assets/LICENSE": "aca6d0ebfbc14cb467904e5a7a21393d",
+"assets/AssetManifest.json": "2efbb41d7877d10aac9d091f58ccd7b9",
+"assets/FontManifest.json": "01700ba55b08a6141f33e168c4a6c22f",
+"assets/packages/cupertino_icons/assets/CupertinoIcons.ttf": "115e937bb829a890521f72d2e664b632",
+"assets/fonts/MaterialIcons-Regular.ttf": "56d3ffdef7a25659eab6a68a3fbfaf16"
 };
 
 self.addEventListener('activate', function (event) {
@@ -32,9 +33,7 @@ self.addEventListener('fetch', function (event) {
         if (response) {
           return response;
         }
-        return fetch(event.request, {
-          credentials: 'include'
-        });
+        return fetch(event.request);
       })
   );
 });